+++ server/plugins/nagioscorepassivecheck/nagioscorepassivecheck.inc.php
@@ -220,10 +220,11 @@ function nrdp_write_check_output_to_cmd($hostname, $servicename, $state, $output
     // Create a temp file to write to
     $tmpname = tempnam($cfg["check_results_dir"], "c");
 
-    // Check if the file is in the check_results_dir
-    if (strpos($tmpname, $cfg["check_results_dir"]) === false) {
+    // Check if the file is in the check_results_dir (or its symlink)
+    if (strpos($tmpname, realpath($cfg["check_results_dir"])) === false) {
+
         unlink($tmpname);
-        _debug("tmpname({$tmpname}) not in cfg[check_results_dir] ({$cfg['check_results_dir']}), bailing");
+        _debug("tmpname({$tmpname}) not in cfg[check_results_dir] ({$cfg['check_results_dir']}), (or a symlink) bailing");
         handle_api_error(ERROR_BAD_CHECK_RESULTS_DIR);
     }